+++ BLI_GA_Test/Models/ConfigModel.cs
@@ -11,13 +11,13 @@ namespace BLI_GA_Test.Models
         public int IndividualListSize { get; set; }
         public int PopulationSize { get; set; }
         public int MaxNumberOfIterations_GA { get; set; }
-        public Dictionary<string, int> ImportanceDegree { get; set; }
-        public Dictionary<string, int> ImportanceDegree_Fitness { get; set; }
+        public Dictionary<string, double> ImportanceDegree { get; set; }
+        public Dictionary<string, double> ImportanceDegree_Fitness { get; set; }
 
         public ConfigModel() 
         {
-            ImportanceDegree = new Dictionary<string, int>();
-            ImportanceDegree_Fitness = new Dictionary<string, int>();
+            ImportanceDegree = new Dictionary<string, double>();
+            ImportanceDegree_Fitness = new Dictionary<string, double>();
         }
     }
 }
